Additional fixes to new time series implementation
* tests passing, except opmode
